blog modal updated

+++ api/src/controllers/blog.js
@@ -28,12 +28,12 @@ module.exports = {
     */
        
         // const data = await Blog.find()
-        // const data = await res.getModelList(Blog, {},["userId","categoryId"])
-        const data = await res.getModelList(Blog, {},[
-            {path:"userId",
-            select:["username","image"]
-            },{path:"categoryId"}])
-
+        const blogFilters = !(req.user?._id)
+        const blogStatus = !(req.user?.isAdmin || req.user?.isStaff )  ? {isDeleted:false} : {}
+        const blogs = await Blog.find({userId: req.user?._id})
+        console.log(blogs);
+     
+        const data = await res.getModelList(Blog, {...blogStatus},["userId","categoryId"])
 
         res.status(200).send({
             error: false,
@@ -81,19 +81,9 @@ module.exports = {
 
 
         const userId = req.user?._id
-
-        const data = await Blog.findOne({ _id: req.params.blogId })
-            .populate({
-                path: 'userId',
-                select: ['username', 'image']
-            })
-            .populate({
-                path: 'comments',
-                populate: {
-                    path: 'userId',
-                    select: ['username', 'image']
-                }
-            });
+        const customFilter = !(req.user?.isAdmin || req.user?.isStaff) ? {isDeleted: false,isPublish: true} : {}
+        const data = await Blog.findOne({ _id: req.params.blogId, ...customFilter }).populate("userId")
+       
         if(!data){
             throw new Error("There is no such a blog, it is removed sorry")
         }
@@ -103,7 +93,7 @@ module.exports = {
              await data.save()
         }
 
-       console.log("countOfViews",data.countOfViews);
+    //    console.log("countOfViews",data.countOfViews);
         res.status(202).send({
             error: false,
             data
@@ -125,8 +115,15 @@ module.exports = {
         }
     */
 
+        const blog = await Blog.findOne({_id: req.params.blogId})
+        if(!(req.user?.isAdmin || req.user?.isStaff ) || (blog?.userId !== req.user?._id)) {
+            res.status(403).send({
+                error: true,
+                message:"You are not the owner of the blog to update it"
+            })
+        }
         const data = await Blog.updateOne({ _id: req.params.blogId}, req.body, { runValidators: true })
-
+       
         res.status(202).send({
             error: false,
             data,
@@ -142,10 +139,25 @@ module.exports = {
         #swagger.summary = "Delete Blog"
     */
 
-        const { deletedCount } = await Blog.deleteOne({ _id: req.params.blogId })
+        // const { deletedCount } = await Blog.deleteOne({ _id: req.params.blogId })
+        const blogs = await Blog.findOne({userId: req.user?._id})
 
-        res.status(deletedCount ? 204 : 404).send({
-            error: !(!!deletedCount)
+        if(!(req.user?.isAdmin || req.user?.isStaff) || req.user?._id !== blogs?.userId){
+           res.status(403).send({
+            message: "You are not the owner of the blog to do this operation"
+           })
+        }
+         if(req.user?.isAdmin || req.user?.isStaff) {
+            const { deletedCount } = await Blog.deleteOne({ _id: req.params.blogId })
+            res.status(deletedCount ? 204 : 404).send({
+                message:"Blog deleted by the authorized person permanently"
+            })
+         }
+         const deletedBlog = await Blog.updateOne({ _id: req.params.blogId },{isDeleted: true})
+
+        res.status(deletedBlog ? 204 : 404).send({
+            // error: !(!!deletedBlog)
+            message: "Deleted successfully"
         })
     },
 }
